zoom, save/load, org. code, resize buttons.
Fix terminology for zoom, as it was in reverse.
Limit zoom in to ¼.
Zoom value is now kept in the xml file.
Save_xml and load_xml as now parameter called filename.
Recognizing code.
Buttons is now resized for the text and font (who says custom gui's
can't do that?).